Integer conversion

+++ includes.py
@@ -24,7 +24,7 @@ def getSummaryList():
 	return [i.attributes['comment'].value for i in changes]
 def getCounts():
 	sums = getSummaryList()
-	revert = round((len(([i for i in sums if classifyEdit(i)]))+1)/5.0)
-	total = round((len(sums)+1)/5.0)
+	revert = int(round((len(([i for i in sums if classifyEdit(i)]))+1)/5.0))
+	total = int(round((len(sums)+1)/5.0))
 	return (total, revert)
